Add SearchService and related resources
Added a new microservice `SearchService` for car auction application. This service will handle the search functionality for the car auctions platform and also manage API endpoints relative to the `Search` model.

The `SearchService` includes configurations files, a new SearchController.cs, an associated Program.cs and the project build settings in a new SearchService.csproj file. A new Docker container was also added for `mongodb` in the `docker-compose.yml` file to store the search data. Added seed data for MongoDb entity `Item` and a `DbInitializer` for seeding data.

These changes are necessary for the development of the search function in the car auctions platform.