Remove `InitialCreate` migration and replace with `Initial` for AuctionService.
This commit removes the `InitialCreate` migration and replaces it with a more comprehensive `Initial` migration. The new `Initial` migration includes tables for `InboxState`, `OutboxMessage`, and `OutboxState`, which were not previously present. The reason for this change is to incorporate MassTransit's outbox and inbox features which help ensure reliable message delivery between microservices in the application. Other changes in this commit include updates in project setting and codes which relate to MassTransit's configuration.